refactor: split column_mapper_step.dart into 4 themed extensions
The _ColumnMapperStep extension on _ImportScreenState was 1493 LoC.
Split it into 4 extensions, each in its own part file of
import_screen.dart (Dart allows multiple extensions on the same type,
so behavior is identical):

- column_mapper_step.dart (kept): _buildColumnMapper +
  _buildSingleAssetPicker + _showCreateEmptyAssetDialog +
  _buildInlineAccountSelector (311 LoC, was 1493)
- mapping_content.dart: _buildMappingContent + _buildMappingRow
  (518 LoC)
- mode_sections.dart: balance / type detection / fee sections
  (363 LoC)
- amount_formula.dart: amount mode buttons / formula row /
  balance preview (313 LoC)

Distribution across lib/:
  max:              1493 → 1255
  files > 1000 LoC: 6 → 5
  avg LoC/file:     329 → 320

All four suites green: analyze clean, 778 unit, integration +4, live
data fetch +1.
